Refactoring: extract e2e infra to the superclass
